main.py
#!/usr/bin/env python3
"""
Kayaking Conditions Notifier for Morro Bay
Main script that orchestrates the notification service.
"""

import os
import sys
from datetime import datetime, timedelta
import pytz
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

def main():
    """Main function to run the kayaking conditions check and notification."""
    try:
        # TODO: Implement main logic
        pass
    except Exception as e:
        print(f"Error: {str(e)}", file=sys.stderr)
        sys.exit(1)

if __name__ == "__main__":
    main()
