requirements.txt
python-dotenv>=1.0.0
requests>=2.31.0
pyyaml>=6.0.1
openai>=1.6.1
anthropic>=0.8.1
pushover-complete>=1.1.1
schedule>=1.2.1
pytz>=2023.3
python-dateutil>=2.8.2
